requirements.txt
Django==2.2.4
psycopg2-binary==2.8.3
django-countries==5.4
djangorestframework==3.10.3
django-cors-headers==3.1.0